tx: lunatix 0.13.0 - integration
It compiles.
Not tested yet.

--

+++ src/entities/TargetShooter.hpp
@@ -39,7 +39,7 @@ class TargetShooter : public Enemy
 protected:
 
     unsigned int id;
-    int vel;
+    Float vel;
 
 public:
